+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added the ability to select a cell by pressing Escape on an auto focused button or anchor element (terra-data-grid only).
+
 ## 5.16.0 - (March 25, 2024)
 
 * Fixed